Move constraint capture cluster to its own file
